+++ src/pages/Profile.js
@@ -2,6 +2,7 @@ import React from "react";
 import LoginForm from "../components/loginForm/LoginForm";
 import RegistrationForm from "../components/registrationForm/RegistrationForm";
 import Menu from "../components/menu/Menu";
+import ProfileCard from '../components/profileCard'
 import { Avatar } from 'antd';
 import { userIsAuthenticated } from "../redux/HOCs";
 import { UserOutlined } from '@ant-design/icons';
@@ -9,6 +10,10 @@ import { Layout } from 'antd';
 
 
 class Profile extends React.Component {
+  getUsername() {
+    let path = window.location.pathname.split('/')
+    return path[2]
+  }
   render() {
 
     const { Header, Content, Footer} = Layout;
@@ -23,57 +28,10 @@ class Profile extends React.Component {
       fontFamily: 'fantasy',
     }
 
-    const content = {
-      backgroundColor: '#fffb8f',
-      padding: '50px 0'
-    }
-
-    const card = {
-      boxShadow: '0 4px 8px 0 rgba(0, 0, 0, 0.2)',
-      maxWidth: '300px',
-      fontFamily: 'Arial',
-      margin: 'auto',
-      textAlign: 'center',
-      fontFamily: 'arial',
-      backgroundColor: 'ffffff',
-
-    }
-
-    const avatar = {
-      backgroundColor: '#e6f7ff',
-      height: '110px'
-      
-    }
-
-    const user = {
-      username: 'Username',
-      displayName: 'Display Name',
-      about: 'Kenzie Student',
-      createdAt: 'Mar-09-2020',
-      updatedAt: 'Today',
-      pictureLocation: 'sample location',
-      googleId: 'google',
-     
-
-    }
-
-    const button = {
-      border: 'none',
-      outline: '0',
-      display: 'inline-block',
-      padding: '8px',
-      color: 'white',
-      backgroundColor: '#000',
-      textAlign: 'center',
-      cursor: 'pointer',
-      width: '94%',
-      height: '60px',
-      fontSize: '18px',
-    }
-
       const footer = {
         backgroundColor: '#73d13d',
       }
+
 
     return (
       <>
@@ -81,19 +39,7 @@ class Profile extends React.Component {
           <Header style={header} className="header">
             Rioters Reinvented
           </Header>
-          <Content style={content} className="content">
-            <div style={card} className="Profile">
-              <div style={avatar} className="avatar">
-                <Avatar shape="square" size={130} icon={<UserOutlined />} />
-               </div>
-              <p>{user.username}</p>
-              <p>{user.about}</p>
-              <p>{user.createdAt}</p>
-              <div style={button} className="Button">
-                <button>Message</button>
-              </div>
-            </div>
-          </Content>
+          <ProfileCard username={this.getUsername()}/>
           <Footer style={footer} className="footer">
           <Menu isAuthenticated={this.props.isAuthenticated} />
             <p>Menu</p>
